Leaf classes for air unit testing
Currently only testing with NASA9. NASA7 will come with more parsing
capabilities deployed for the Antioch interface. Note that the tolerance
is very loose for Cantera - can't really explain that at this point.
This was the best I could come up with the for moment to reuse most of
the code and make it easy to add a different mechanism, e.g. Ozone.